Revert "Merge branch 'uw/master' into dev/V7_7-gt2472-classad-64-bit-int"
Still seeing build and test failures -- let's revert at try again.

This reverts commit 1b0ba1dd647ccb62844b830ab91c1d5c2dfc7c29, reversing
changes made to b7a4eda53c0fde16ed29926b571ee30dd0140197.

This reverts the following:

git revert -n cf6f26fa 1845eb59268 58f0a86 7dfd5cca 8188696 e775235 a8e89dc
git revert -n -m 2 1b0ba1dd

+++ src/condor_utils/param_info.cpp
@@ -85,10 +85,10 @@ param_default_string(const char* param)
 	return ret;
 }
 
-long long
+int
 param_default_integer(const char* param, int* valid) {
 	const param_info_t* p;
-    long long ret = 0;
+	int ret = 0;
 
 	param_info_init();
 
@@ -133,7 +133,7 @@ param_default_double(const char* param, int* valid) {
 }
 
 int
-param_range_integer(const char* param, long long* min, long long* max) {
+param_range_integer(const char* param, int* min, int* max) {
 
 	const param_info_t* p;
